+++ aegislab/src/internal/chaosengine/guidedcli/builders.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	"aegis/internal/chaosengine/handler"
-	"aegis/internal/chaosengine/resourcelookup"
+	"aegis/platform/k8s/resourcelookup"
 	"aegis/platform/systemconfig"
 )
