+++ aegislab/src/internal/chaosengine/guidedcli/preview.go
@@ -1,7 +1,7 @@
 package guidedcli
 
 import (
-	"aegis/internal/chaosengine/resourcelookup"
+	"aegis/platform/k8s/resourcelookup"
 	"aegis/platform/systemconfig"
 )
