+++ aegislab/src/internal/chaosengine/guidedcli/resolver.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 
 	"aegis/internal/chaosengine/handler"
-	"aegis/internal/chaosengine/resourcelookup"
+	"aegis/platform/k8s/resourcelookup"
 	"aegis/platform/systemconfig"
 )
